datetime test case

+++ tests/libs/libc/datetimetest.cpp
@@ -6,6 +6,6 @@ using namespace Syscall;
 
 Time dateTime::current() {
     Time result;
-    doSyscall(SYSCALL_GET_DATETIME, (uint8_t)&result);
+    doSyscall(SYSCALL_GET_DATETIME, (uint32_t)&result);
     return result;
 }
